move animation component into components folder
