--- src/animations/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  get AnimatedImage() { return require('./animatedImage').default; },
-  get AnimatedScanner() { return require('./animatedScanner').default; },
-  get ProgressBar() { return require('./progressBar').default; },
-};
