+++ src/components/animatedScanner/index.js
