+++ src/components/progressBar/index.js
